refactor(pty): split TerminalProcess into focused modules
- Extract identityDebug.ts (dedupes ProcessDetector)
- Extract terminalTitle.ts (computeDefaultTitle / getLiveAgentId, with renderer-lockstep comment)
- Extract terminalSerialization.ts (pure functions for serialize / serializeAsync / serializeForPersistence)
- Extract ForegroundProcessGroupProbe.ts (stale-while-revalidate ps probe)
- Extract TerminalExitObservers.ts (terminal:exited emission + forensics/agent:completed/fallback subscription)
- Extract TerminalGracefulShutdown.ts (quit-command / session-id capture flow)
- Extract TerminalAgentDetection.ts (detection result -> chrome/title/event fanout)
- Extract TerminalProcessLifecycle.ts (alive/shutting-down/exited/disposed state machine)

TerminalProcess.ts: 2007 -> 1411 lines (596 removed). Public surface, constructor signature, and TerminalProcessCallbacks unchanged. setupPtyHandlers retained as wiring choke point. Lesson #3177 (snapshot flush before kill) preserved.

All 812 pty tests pass; no test file modifications required.
